+++ packages/react/src/components/FormField/src/components/FormFieldInput/formFieldInput.types.ts
@@ -3,8 +3,18 @@ import { InputProperties, InputProps } from "@nimbus-ds/components";
 import { FormFieldProperties } from "../../formField.types";
 
 export type FormFieldInputProperties = Omit<FormFieldProperties, "children"> &
-  Omit<InputProperties, "appearance">;
+  Omit<InputProperties, "appearance"> & {
+    appearance?:
+      | FormFieldProperties["appearance"]
+      | InputProperties["appearance"];
+  };
 
-export type FormFieldInputBaseProps = FormFieldInputProperties &
-  Omit<InputProps, "appearance"> &
-  HTMLAttributes<HTMLElement>;
+export type FormFieldInputBaseProps = Omit<
+  FormFieldInputProperties,
+  "appearance"
+> &
+  Omit<InputProps, "appearance"> & {
+    appearance?:
+      | FormFieldProperties["appearance"]
+      | InputProperties["appearance"];
+  } & HTMLAttributes<HTMLElement>;